SupplyChainModeling

Installation

SupplyChainModeling can be installed using the Julia package manager. From the Julia REPL, type ] to enter the Pkg REPL mode and run

pkg> add SupplyChainModeling

API

SupplyChainModeling.Customer — Type

A customer.

SupplyChainModeling.Demand — Type

The demand a customer has for a product.

SupplyChainModeling.Lane — Type

A transportation lane between two nodes of the supply chain.

SupplyChainModeling.Location — Type

The geographical location of a node of the supply chain. The location is defined by its latitude and longitude.

SupplyChainModeling.Node — Type

A node of the supply chain.

SupplyChainModeling.Plant — Type

A plant.

SupplyChainModeling.Storage — Type

A storage location.

SupplyChainModeling.Supplier — Type

A supplier.

SupplyChainModeling.SupplyChain — Type

The supply chain.

SupplyChainModeling.add_customer! — Method

add_customer!(supply_chain, customer)

Adds a customer to the supply chain.

SupplyChainModeling.add_demand! — Method

add_demand!(supply_chain, customer, product; demand::Array{Float64, 1}, service_level=1.0)

Adds customer demand for a product. The demand is specified for each time period.

The keyword arguments are:

demand: the amount of product demanded for each time period.

service_level: indicates how many lost sales are allowed as a ratio of demand. No demand can be lost if the service level is 1.0 and all demand can be lost if the service level is 0.0.

SupplyChainModeling.add_demand! — Method

add_demand!(supply_chain, demand)

Adds demand to the supply chain.

SupplyChainModeling.add_lane! — Method

add_lane!(supply_chain, lane)

Adds a transportation lane to the supply chain.

SupplyChainModeling.add_plant! — Method

add_plant!(supply_chain, plant)

Adds a plant to the supply chain.

SupplyChainModeling.add_product! — Method

add_product!(plant::Plant, product::Product; bill_of_material::Dict{Product, Float64}, unit_cost, maximum_throughput)

Indicates that a plant can produce a product.

The keyword arguments are:

bill_of_material: the amount of other product needed to produce one unit of the product. This dictionary can be empty if there are no other products needed.

unit_cost: the cost of producing one unit of product.

maximum_throughput: the maximum amount of product that can be produced in a time period.

time: the production lead time.

SupplyChainModeling.add_product! — Method

add_product!(supplier::Supplier, product::Product; unit_cost::Float64, maximum_throughput::Float64)

Indicates that a supplier can provide a product.

The keyword arguments are:

unit_cost: the cost per unit of the product from this supplier.

maximum_throughput: the maximum number of units that can be provided in each time period.

SupplyChainModeling.add_product! — Method

add_product!(supply_chain, product)

Adds a product to the supply chain.

SupplyChainModeling.add_storage! — Method

add_storage!(supply_chain, storage)

Adds a storage location to the supply chain.

SupplyChainModeling.add_supplier! — Method

add_supplier!(supply_chain, supplier)

Adds a supplier to the supply chain.

SupplyChainModeling.can_ship — Method

can_ship(lane::Lane, time::Int)

Checks if units can be send on the lane at a given time.

SupplyChainModeling.get_arrivals — Method

get_arrivals(lane::Lane, destination, time::Int)

Gets the known arrivals.